Move out stuff to point.rs
